fix(langgraph): unwrap ToolMessage before serialization in on_tool_end
LangGraph's on_tool_end receives ToolMessage objects that aren't JSON
serializable, causing _serialize_tool_content to error and mark spans
with StatusCode.ERROR. The LangChain handler already handled this but
the LangGraph handler was missed.

Add _extract_tool_output to unwrap .content from message objects at
both sync and async on_tool_end call sites.
